Move tutorial components to shared-styles
