+++ context/app/static/js/shared-styles/tutorials/TutorialTooltip/style.js
